Revert "Wrapper task"
