+++ docs/tasks/FileSystem.md
@@ -67,6 +67,7 @@ $this->_mkdir('logs');
 ?>
 ```
 
+* `stopOnFail($stop = null)` 
 * `mkdir($dir)` 
 * `touch($file)` 
 * `copy($from, $to, $force = null)` 
@@ -77,8 +78,6 @@ $this->_mkdir('logs');
 * `mirror($from, $to)` 
 * `chgrp($file, $group)` 
 * `chown($file, $user)` 
-
-* `stopOnFail($stop = null)` 
 
 ## FlattenDir
 
@@ -160,6 +159,7 @@ $this->_mirrorDir('dist/config/', 'config/');
 
 ?>
 ```
+
